Move ApplicationFormBase to forms.py
